# [performance](bitmap) Support CRoaring COW feature to reduce bitmap copying