--- a/be/src/common/config.cpp
+++ b/be/src/common/config.cpp
@@ -998,6 +998,9 @@ DEFINE_mInt32(s3_write_buffer_whole_size, "524288000");
 //disable shrink memory by default
 DEFINE_Bool(enable_shrink_memory, "false");
 
+// enable croaring cow
+DEFINE_mBool(enable_croaring_cow, "false");
+
 #ifdef BE_TEST
 // test s3
 DEFINE_String(test_s3_resource, "resource");