--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -1014,6 +1014,9 @@ DECLARE_mInt32(s3_write_buffer_whole_size);
 //enable shrink memory
 DECLARE_Bool(enable_shrink_memory);
 
+// enable croaring cow
+DECLARE_mBool(enable_croaring_cow);
+
 #ifdef BE_TEST
 // test s3
 DECLARE_String(test_s3_resource);